D24: global_rx accumulated covariance in float32; tests for crd/streaming/infer/postfilter
274 tests pass, up from 200.

D24: 3A.5 requires streaming_rx to match global_rx to rtol=1e-5 and explains
why the accumulator must be float64 -- "float32 co-moment accumulation loses
precision over 20000+ pixels and quietly biases the covariance." That warning
was written into the streaming spec and never applied to the reference
implementation streaming is validated against. Measured gaps were 8.5e-4
(Indian Pines) and 4.3e-2 (abu-urban-3), i.e. ~4300x looser than spec.

The reference was the imprecise side: global_rx did a float32 matmul over
10000+ pixels, max rel error 1.15e-4 vs a float64 version of its own formula.
streaming_rx was already float64 as the spec demanded, so the two disagreed
exactly to the extent the reference was wrong. Fixed; equivalence is now
bit-identical on Indian Pines and abu-urban-3, 1.2e-7 on abu-beach-1. ABU AUC
moves by <=0.0007.

Worth recording despite the one-line fix: the natural response to a failing
tolerance is to loosen it, and rtol=0.1 would have been green and wrong. A
tolerance is a claim about two implementations; when it fails the reference is
a suspect too. Several more cross-implementation checks land in Phase 5.

Also corrects 3B.7's worked example (found by execution): 0.5*scene_px is 2048
at 64x64 and 7200 at 120x120, both above the base max_area_px=2000, so the
per-scene term is a no-op at exactly the two sizes named to demonstrate it. It
binds only below ~4000 scene px. The rule is right, the illustration was not --
same class as D20's slip.

Training complete: unet_pretext, 40 epochs, best val 0.1243 @ epoch 27, 5.1h
on the GTX 1650.

+++ anomaly/rx.py
@@ -14,7 +14,15 @@ def global_rx(cube: np.ndarray, *, reg: float = 1e-6) -> np.ndarray:
     Uses scipy.linalg.cho_factor / cho_solve -- never an explicit inverse.
     """
     h, w, b = cube.shape
-    flat = cube.reshape(-1, b)
+    # float64 accumulation (D24). The cube is float32, and a float32
+    # `centered.T @ centered` over 10 000+ pixels loses precision and quietly
+    # biases the covariance -- measured max relative error 1.15e-4 against a
+    # float64 reference on abu-urban-3. This is the exact hazard PLAN.md
+    # §3A.5 writes into the STREAMING spec ("float32 co-moment accumulation
+    # loses precision over 20 000+ pixels and quietly biases the
+    # covariance"); the warning was simply never applied to the reference
+    # implementation it was meant to protect.
+    flat = cube.reshape(-1, b).astype(np.float64, copy=False)
     valid = ~np.any(np.isnan(flat), axis=-1)
 
     mu = flat[valid].mean(axis=0)
